+++ singularity_sporg/src/project_settings.rs
